+++ src/internal/kopia/merge_details_test.go
@@ -12,6 +12,18 @@ import (
 	"github.com/alcionai/corso/src/pkg/path"
 )
 
+type mockLocationIDer struct {
+	pb *path.Builder
+}
+
+func (ul mockLocationIDer) ID() *path.Builder {
+	return ul.pb
+}
+
+func (ul mockLocationIDer) InDetails() *path.Builder {
+	return ul.pb
+}
+
 type DetailsMergeInfoerUnitSuite struct {
 	tester.Suite
 }
@@ -20,11 +32,9 @@ func TestDetailsMergeInfoerUnitSuite(t *testing.T) {
 	suite.Run(t, &DetailsMergeInfoerUnitSuite{Suite: tester.NewUnitSuite(t)})
 }
 
-// TestRepoRefs is a basic sanity test to ensure lookups are working properly
-// for stored RepoRefs.
-func (suite *DetailsMergeInfoerUnitSuite) TestRepoRefs() {
+func (suite *DetailsMergeInfoerUnitSuite) TestAddRepoRef_DuplicateFails() {
 	t := suite.T()
-	oldRef := makePath(
+	oldRef1 := makePath(
 		t,
 		[]string{
 			testTenant,
@@ -33,29 +43,142 @@ func (suite *DetailsMergeInfoerUnitSuite) TestRepoRefs() {
 			category,
 			"folder1",
 		},
-		false).ToBuilder()
-	newRef := makePath(
+		false)
+
+	dm := newMergeDetails()
+
+	err := dm.addRepoRef(oldRef1.ToBuilder(), oldRef1, nil)
+	require.NoError(t, err, clues.ToCore(err))
+
+	err = dm.addRepoRef(oldRef1.ToBuilder(), oldRef1, nil)
+	require.Error(t, err, clues.ToCore(err))
+}
+
+// TestRepoRefs is a basic sanity test to ensure lookups are working properly
+// for stored RepoRefs.
+func (suite *DetailsMergeInfoerUnitSuite) TestGetNewPathRefs() {
+	t := suite.T()
+	oldRef1 := makePath(
 		t,
 		[]string{
 			testTenant,
 			service,
 			testUser,
 			category,
+			"folder1",
+		},
+		false)
+	oldRef2 := makePath(
+		t,
+		[]string{
+			testTenant,
+			service,
+			testUser,
+			category,
+			"folder1",
 			"folder2",
 		},
 		false)
+	newRef1 := makePath(
+		t,
+		[]string{
+			testTenant,
+			service,
+			testUser,
+			category,
+			"folder3",
+		},
+		false)
+	newRef2 := makePath(
+		t,
+		[]string{
+			testTenant,
+			service,
+			testUser,
+			category,
+			"folder3",
+			"folder2",
+		},
+		false)
+	newLoc1 := path.Builder{}.Append(newRef1.Folders()...)
+	newLoc2 := path.Builder{}.Append(newRef2.Folders()...)
+	oldLoc1 := path.Builder{}.Append(oldRef1.Folders()...)
+	oldLoc2 := path.Builder{}.Append(oldRef2.Folders()...)
+
+	searchLoc1 := mockLocationIDer{oldLoc1}
+	searchLoc2 := mockLocationIDer{oldLoc2}
 
 	dm := newMergeDetails()
 
-	err := dm.addRepoRef(oldRef, newRef)
+	err := dm.addRepoRef(oldRef1.ToBuilder(), newRef1, newLoc1)
 	require.NoError(t, err, clues.ToCore(err))
 
-	got := dm.GetNewRepoRef(oldRef)
-	require.NotNil(t, got)
-	assert.Equal(t, newRef.String(), got.String())
+	err = dm.addRepoRef(oldRef2.ToBuilder(), newRef2, nil)
+	require.NoError(t, err, clues.ToCore(err))
 
-	got = dm.GetNewRepoRef(newRef.ToBuilder())
-	assert.Nil(t, got)
+	// Add prefix matcher entry.
+	err = dm.addLocation(searchLoc1, newLoc1)
+	require.NoError(t, err, clues.ToCore(err))
+
+	table := []struct {
+		name        string
+		searchRef   *path.Builder
+		searchLoc   mockLocationIDer
+		errCheck    require.ErrorAssertionFunc
+		expectedRef path.Path
+		expectedLoc *path.Builder
+	}{
+		{
+			name:        "Exact Match With Loc",
+			searchRef:   oldRef1.ToBuilder(),
+			searchLoc:   searchLoc1,
+			errCheck:    require.NoError,
+			expectedRef: newRef1,
+			expectedLoc: newLoc1,
+		},
+		{
+			name:        "Exact Match Without Loc",
+			searchRef:   oldRef1.ToBuilder(),
+			errCheck:    require.NoError,
+			expectedRef: newRef1,
+			expectedLoc: newLoc1,
+		},
+		{
+			name:        "Prefix Match",
+			searchRef:   oldRef2.ToBuilder(),
+			searchLoc:   searchLoc2,
+			errCheck:    require.NoError,
+			expectedRef: newRef2,
+			expectedLoc: newLoc2,
+		},
+		{
+			name:      "Would Be Prefix Match Without Old Loc Errors",
+			searchRef: oldRef2.ToBuilder(),
+			errCheck:  require.Error,
+		},
+		{
+			name:      "Not Found With Old Loc",
+			searchRef: newRef1.ToBuilder(),
+			searchLoc: searchLoc2,
+			errCheck:  require.NoError,
+		},
+		{
+			name:      "Not Found Without Old Loc",
+			searchRef: newRef1.ToBuilder(),
+			errCheck:  require.NoError,
+		},
+	}
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
+
+			newRef, newLoc, err := dm.GetNewPathRefs(test.searchRef, test.searchLoc)
+			test.errCheck(t, err, clues.ToCore(err))
+
+			assert.Equal(t, test.expectedRef, newRef, "RepoRef")
+			assert.Equal(t, test.expectedLoc, newLoc, "LocationRef")
+		})
+	}
 }
 
 type LocationPrefixMatcherUnitSuite struct {
@@ -66,14 +189,9 @@ func TestLocationPrefixMatcherUnitSuite(t *testing.T) {
 	suite.Run(t, &LocationPrefixMatcherUnitSuite{Suite: tester.NewUnitSuite(t)})
 }
 
-type inputData struct {
-	repoRef path.Path
-	locRef  *path.Builder
-}
-
 func (suite *LocationPrefixMatcherUnitSuite) TestAdd_Twice_Fails() {
 	t := suite.T()
-	p := makePath(
+	p := mockLocationIDer{makePath(
 		t,
 		[]string{
 			testTenant,
@@ -82,7 +200,7 @@ func (suite *LocationPrefixMatcherUnitSuite) TestAdd_Twice_Fails() {
 			category,
 			"folder1",
 		},
-		false).ToBuilder()
+		false).ToBuilder()}
 	loc1 := path.Builder{}.Append("folder1")
 	loc2 := path.Builder{}.Append("folder2")
 
@@ -96,93 +214,53 @@ func (suite *LocationPrefixMatcherUnitSuite) TestAdd_Twice_Fails() {
 }
 
 func (suite *LocationPrefixMatcherUnitSuite) TestAdd_And_Match() {
-	p1 := makePath(
-		suite.T(),
-		[]string{
-			testTenant,
-			service,
-			testUser,
-			category,
-			"folder1",
-		},
-		false)
+	loc1 := mockLocationIDer{path.Builder{}.Append("folder1")}
+	loc2 := mockLocationIDer{loc1.InDetails().Append("folder2")}
+	loc3 := mockLocationIDer{path.Builder{}.Append("foo")}
 
-	loc1 := path.Builder{}.Append("folder1")
-	p1Parent, err := p1.Dir()
+	res1 := mockLocationIDer{path.Builder{}.Append("1")}
+
+	lpm := newLocationPrefixMatcher()
+
+	err := lpm.add(loc1, res1.InDetails())
 	require.NoError(suite.T(), err, clues.ToCore(err))
-
-	p2 := makePath(
-		suite.T(),
-		[]string{
-			testTenant,
-			service,
-			testUser,
-			category,
-			"folder2",
-		},
-		false)
 
 	table := []struct {
 		name      string
-		inputs    []inputData
-		searchKey string
-		check     require.ValueAssertionFunc
-		expected  *path.Builder
+		searchKey mockLocationIDer
+		found     bool
 	}{
 		{
-			name: "Exact Match",
-			inputs: []inputData{
-				{
-					repoRef: p1,
-					locRef:  loc1,
-				},
-			},
-			searchKey: p1.String(),
-			check:     require.NotNil,
-			expected:  loc1,
+			name:      "Exact Match",
+			searchKey: loc1,
+			found:     true,
 		},
 		{
-			name: "No Match",
-			inputs: []inputData{
-				{
-					repoRef: p1,
-					locRef:  loc1,
-				},
-			},
-			searchKey: p2.String(),
-			check:     require.Nil,
+			name:      "No Match",
+			searchKey: loc3,
 		},
 		{
-			name: "No Prefix Match",
-			inputs: []inputData{
-				{
-					repoRef: p1Parent,
-					locRef:  loc1,
-				},
-			},
-			searchKey: p1.String(),
-			check:     require.Nil,
+			name:      "Prefix Match",
+			searchKey: loc2,
+			found:     true,
 		},
 	}
 
 	for _, test := range table {
 		suite.Run(test.name, func() {
 			t := suite.T()
-			lpm := newLocationPrefixMatcher()
 
-			for _, input := range test.inputs {
-				err := lpm.add(input.repoRef.ToBuilder(), input.locRef)
-				require.NoError(t, err, clues.ToCore(err))
-			}
+			prefixes := lpm.longestPrefix(test.searchKey)
 
-			loc := lpm.longestPrefix(test.searchKey)
-			test.check(t, loc)
+			if !test.found {
+				assert.Nil(t, prefixes.oldLoc)
+				assert.Nil(t, prefixes.newLoc)
 
-			if loc == nil {
 				return
 			}
 
-			assert.Equal(t, test.expected.String(), loc.String())
+			assert.Equal(t, loc1.InDetails(), prefixes.oldLoc, "old prefix")
+			assert.Equal(t, res1.InDetails(), prefixes.newLoc, "new prefix")
 		})
 	}
 }