+++ src/internal/operations/backup_integration_test.go
@@ -844,9 +844,11 @@ func (suite *BackupOpIntegrationSuite) TestBackup_Run_exchangeIncrementals() {
 
 	// verify test data was populated, and track it for comparisons
 	for category, gen := range dataset {
+		ss := selectors.Selector{}.SetDiscreteOwnerIDName(suite.user, suite.user)
+
 		qp := graph.QueryParams{
 			Category:      category,
-			ResourceOwner: suite.user,
+			ResourceOwner: ss,
 			Credentials:   m365,
 		}
 		cr, err := exchange.PopulateExchangeContainerResolver(ctx, qp, fault.New(true))
@@ -958,9 +960,11 @@ func (suite *BackupOpIntegrationSuite) TestBackup_Run_exchangeIncrementals() {
 						version.Backup,
 						gen.dbf)
 
+					ss := selectors.Selector{}.SetDiscreteOwnerIDName(suite.user, suite.user)
+
 					qp := graph.QueryParams{
 						Category:      category,
-						ResourceOwner: suite.user,
+						ResourceOwner: ss,
 						Credentials:   m365,
 					}
 					cr, err := exchange.PopulateExchangeContainerResolver(ctx, qp, fault.New(true))